+++ src/platform.test.ts
@@ -15,6 +15,7 @@ import {
   IdentifyCluster,
   Matterbridge,
   MatterbridgeDevice,
+  MatterbridgeEndpoint,
   OnOffCluster,
   PlatformConfig,
   PowerSourceCluster,
@@ -102,13 +103,39 @@ describe('ShellyPlatform', () => {
   beforeAll(() => {
     // Creates the mocks for Matterbridge, AnsiLogger, and PlatformConfig
     mockMatterbridge = {
-      matterbridgeDirectory: 'temp',
-      matterbridgePluginDirectory: 'temp',
-      systemInformation: { ipv4Address: undefined },
-      matterbridgeVersion: '1.6.6',
-      addBridgedDevice: jest.fn(),
-      removeBridgedDevice: jest.fn(),
-      removeAllBridgedDevices: jest.fn(),
+      matterbridgeDirectory: './jest/matterbridge',
+      matterbridgePluginDirectory: './jest/plugins',
+      systemInformation: { ipv4Address: undefined, osRelease: 'xx.xx.xx.xx.xx.xx', nodeVersion: '22.1.10' },
+      matterbridgeVersion: '1.7.1',
+      edge: false,
+      log: mockLog,
+      getDevices: jest.fn(() => {
+        // console.log('getDevices called');
+        return [];
+      }),
+      getPlugins: jest.fn(() => {
+        // console.log('getDevices called');
+        return [];
+      }),
+      addBridgedDevice: jest.fn(async (pluginName: string, device: MatterbridgeDevice) => {
+        // console.log('addBridgedDevice called');
+      }),
+      addBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
+        // console.log('addBridgedEndpoint called');
+        // await aggregator.add(device);
+      }),
+      removeBridgedDevice: jest.fn(async (pluginName: string, device: MatterbridgeDevice) => {
+        // console.log('removeBridgedDevice called');
+      }),
+      removeBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
+        // console.log('removeBridgedEndpoint called');
+      }),
+      removeAllBridgedDevices: jest.fn(async (pluginName: string) => {
+        // console.log('removeAllBridgedDevices called');
+      }),
+      removeAllBridgedEndpoints: jest.fn(async (pluginName: string) => {
+        // console.log('removeAllBridgedEndpoints called');
+      }),
     } as unknown as Matterbridge;
     mockLog = {
       fatal: jest.fn((message) => {
@@ -133,11 +160,12 @@ describe('ShellyPlatform', () => {
     mockConfig = {
       'name': 'matterbridge-shelly',
       'type': 'DynamicPlatform',
-      'version': '1.0.0',
+      'version': '1.1.2',
       'username': 'admin',
       'password': 'tango',
       'exposeSwitch': 'outlet',
       'exposeInput': 'contact',
+      'exposeInputEvent': 'momentary',
       'exposePowerMeter': 'matter13',
       'blackList': [],
       'whiteList': [],
@@ -324,7 +352,7 @@ describe('ShellyPlatform', () => {
   it('should throw because of version', () => {
     mockMatterbridge.matterbridgeVersion = '1.5.4';
     expect(() => new ShellyPlatform(mockMatterbridge, mockLog, mockConfig)).toThrow();
-    mockMatterbridge.matterbridgeVersion = '1.6.6';
+    mockMatterbridge.matterbridgeVersion = '1.7.1';
   });
 
   it('should call onStart with reason and start mDNS', async () => {